requirements.txt
urllib3
kafka-python
psycopg2-binary
requests
toml